Refactor and testing ReadJson
Reading TestReadJson and check if the bytes is correct from my function
and the readjson code.
Probably not the best testing...
But I need it some ways to be able to test the other parts to verify
the json data that i get and see that the structs is correct.

Still need to solve pushing errors down to main.